lesson-2/requirements.txt
python-dotenv
langchain_core
langchain_openai
langchain_community
langgraph
